Rename rbf.rs to fingerprint.rs
