+++ src/crates/heuristics/src/ast/fingerprint.rs
